[CLEANUP beta] Remove keys.js
